Made test for submitters
