+++ tests/unit/bout_runners/bookkeeper/__init__.py
@@ -1,1 +1,1 @@
-"""Package containing the unit database tests for bout_runners."""
+"""Package containing the unit test for the bokkeeper package."""
