+++ tests/unit/bout_runners/submitter/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the unit test for the submitter package."""
